+++ packages/expo-sensors/plugin/src/withSensors.ts
@@ -8,10 +8,11 @@ import {
 const pkg = require('expo-sensors/package.json');
 const MOTION_USAGE = 'Allow $(PRODUCT_NAME) to access your device motion';
 
-const withSensors: ConfigPlugin<{ motionPermission?: string | false } | void> = (
-  config,
-  { motionPermission } = {}
-) => {
+export type WithSensorsProps = {
+  motionPermission?: string | false;
+};
+
+const withSensors: ConfigPlugin<WithSensorsProps | void> = (config, { motionPermission } = {}) => {
   if (motionPermission === false) {
     config = withPodfileProperties(config, (config) => {
       config.modResults.MOTION_PERMISSION = 'false';