+++ packages/expo-sqlite/CHANGELOG.md
@@ -12,6 +12,8 @@
 
 ### 💡 Others
 
+- export config plugin types ([#32786](https://github.com/expo/expo/pull/32786) by [@leonhh](https://github.com/leonhh))
+
 ## 15.0.2 — 2024-11-07
 
 ### 💡 Others